Restore random_test
